Trait glsl_layout::AsStd140

pub unsafe trait AsStd140 {
    type Align: Copy + Default;
    type Std140: Std140 + ?Sized;
    fn std140(&self) -> Self::Std140
    where
        Self::Std140: Sized;
}

Structure to transform data from rust's structure to the raw data ready to upload to UBO. Users should prefer to use derive(AsStd140) instead of implementing this manually.

Associated Types

type Align: Copy + Default

ZST that enforces alignment required for this type.

type Std140: Std140 + ?Sized

Type that contain same data with memory layout matching glsl's layout(std140).

Loading content...
Required methods

fn std140(&self) -> Self::Std140 where
Self::Std140: Sized,

Get aligned data from structure.

Loading content...