Trait glsl_layout::Uniform

pub trait Uniform: Copy {
    type Align: Copy + Default;
    type Std140: Std140;
    fn std140(&self) -> Self::Std140;
}

Structure to transform data from rust’s structure to the raw data ready to upload to UBO. Users should prepfer to use derive(Uniform) instead of implementing this manually.

Associated Types

type Align: Copy + Default

ZST that enforces alignment required for this type.

type Std140: Std140

Type that contain same data with memory layout matching glsl’s layout(std140).

Loading content...
Required methods

fn std140(&self) -> Self::Std140

Get aligned data from structure.

Loading content...