Trait glsl_layout::Uniform

pub trait Uniform: Copy {
    type Align: Copy;
    type Std140: Std140;
    fn align() -> Self::Align;

    fn std140(&self) -> Self::Std140;
}

Structure to transform data from rust's structure to the raw data ready to upload to UBO. Users should prepfer to use derive(Uniform) instead of implementing this manually.

Associated Types

type Align: Copy

ZST that enforces alignment required for this type.

type Std140: Std140

Type that contain same data with memory layout matching glsl's layout(std140).

Required Methods

fn align() -> Self::Align

The value of the Align.

fn std140(&self) -> Self::Std140

Get aligned data from structure.